requirements.txt
# Web Framework
streamlit==1.31.0

# PDF Processing
pymupdf==1.23.8

# Image Processing
pillow==10.2.0

# Numerical Computing
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
torch==2.1.2
transformers==4.36.2

# Environment Variables
python-dotenv==1.0.0

# LangChain
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-google-genai==0.0.6
langchain-text-splitters==0.0.1

# Vector Store
faiss-cpu==1.7.4

# Retry Logic
tenacity==8.2.3